+++ app/models/__init__.py
@@ -1,0 +1,12 @@
+from app.models.associations import joined_sessions, saved_sessions
+from app.models.user import User
+from app.models.studybuddy import StudySession, SessionMessage
+
+
+__all__ = [
+    "joined_sessions",
+    "saved_sessions",
+    "User",
+    "StudySession",
+    "SessionMessage",
+]
